+++ pallets/common/src/erc.rs
@@ -420,6 +420,7 @@ where
 	///
 	/// @param user account to verify
 	/// @return "true" if account is the owner or admin
+	#[solidity(rename_selector = "verifyOwnerOrAdmin")]
 	fn verify_owner_or_admin_substrate(&self, user: uint256) -> Result<bool> {
 		let user = convert_substrate_address_to_cross_account_id::<T>(user);
 		Ok(self
@@ -455,6 +456,7 @@ where
 	///
 	/// @dev Owner can be changed only by current owner
 	/// @param newOwner new owner substrate account
+	#[solidity(rename_selector = "changeOwner")]
 	fn change_owner_substrate(&mut self, caller: caller, new_owner: uint256) -> Result<void> {
 		let caller = T::CrossAccountId::from_eth(caller);
 		let new_owner = convert_substrate_address_to_cross_account_id::<T>(new_owner);